api_gateway/requirements.txt
Flask==3.1.0
Werkzeug==3.1.3
flask-jwt-extended>=4.7.1
gunicorn==23.0.0
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.36
Flask-CORS==5.0.0
pydantic==2.10.3
